day30/day30_performance_monitoring/requirements.txt
asyncio==3.4.3
aiohttp_cors==0.7.0
aiohttp==3.9.5
aiofiles==23.2.1
psutil==5.9.8
prometheus-client==0.20.0
matplotlib==3.8.4
plotly==5.19.0
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
jinja2==3.1.4
websockets==12.0
uvloop==0.19.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
flask==3.0.3
bokeh==3.4.1
